--- ecc/blocks/tag-manager-component/controller.sample.js
@@ -1,20 +1,0 @@
-/* eslint-disable no-unused-vars */
-export function onSubmit(_component, _props) {
-  // Do nothing
-}
-
-export async function onPayloadUpdate(_component, _props) {
-  // Do nothing
-}
-
-export async function onRespUpdate(_component, _props) {
-  // Do nothing
-}
-
-export default function init(_component, _props) {
-  // Do nothing
-}
-
-export function onTargetUpdate(_component, _props) {
-  // Do nothing
-}
